+++ src/config.rs
@@ -42,6 +42,7 @@ pub struct RtcConfig {
     pub(crate) codec_config: CodecConfig,
     pub(crate) exts: ExtensionMap,
     pub(crate) stats_interval: Option<Duration>,
+    pub(crate) intervals: RtcpReportIntervals,
     pub(crate) bwe_config: Option<BweConfig>,
     pub(crate) reordering_size_audio: usize,
     pub(crate) reordering_size_video: usize,
@@ -58,6 +59,18 @@ pub struct RtcConfig {
 #[derive(Debug, Clone)]
 pub(crate) struct BweConfig {
     pub(crate) initial_bitrate: Bitrate,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) struct RtcpReportIntervals {
+    pub(crate) audio: Duration,
+    pub(crate) video: Duration,
+}
+
+impl RtcpReportIntervals {
+    pub(crate) fn for_audio(self, audio: bool) -> Duration {
+        if audio { self.audio } else { self.video }
+    }
 }
 
 impl RtcConfig {
@@ -383,6 +396,42 @@ impl RtcConfig {
     /// ```
     pub fn stats_interval(&self) -> Option<Duration> {
         self.stats_interval
+    }
+
+    /// Sets the interval between RTCP sender/receiver reports for audio streams.
+    ///
+    /// Defaults to 5 seconds.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `interval` is zero.
+    pub fn set_rtcp_report_interval_audio(mut self, interval: Duration) -> Self {
+        assert!(!interval.is_zero());
+        self.intervals.audio = interval;
+        self
+    }
+
+    /// Returns the interval between RTCP sender/receiver reports for audio streams.
+    pub fn rtcp_report_interval_audio(&self) -> Duration {
+        self.intervals.audio
+    }
+
+    /// Sets the interval between RTCP sender/receiver reports for video streams.
+    ///
+    /// Defaults to 1 second.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `interval` is zero.
+    pub fn set_rtcp_report_interval_video(mut self, interval: Duration) -> Self {
+        assert!(!interval.is_zero());
+        self.intervals.video = interval;
+        self
+    }
+
+    /// Returns the interval between RTCP sender/receiver reports for video streams.
+    pub fn rtcp_report_interval_video(&self) -> Duration {
+        self.intervals.video
     }
 
     /// Enables estimation of available bandwidth (BWE).
@@ -712,6 +761,10 @@ impl Default for RtcConfig {
             codec_config: CodecConfig::new_with_defaults(),
             exts: ExtensionMap::standard(),
             stats_interval: None,
+            intervals: RtcpReportIntervals {
+                audio: Duration::from_secs(5),
+                video: Duration::from_secs(1),
+            },
             bwe_config: None,
             reordering_size_audio: 15,
             reordering_size_video: 30,